adding in all types for all databases

+++ src/main/java/synapticloop/h2zero/model/field/DecimalField.java
@@ -35,12 +35,12 @@ public class DecimalField extends BaseField {
 
 	@Override
 	public String getJavaType() {
-		return("Float");
+		return("Double");
 	}
 
 	@Override
 	public String getSqlJavaType() {
-		return("Float");
+		return("Double");
 	}
 
 	@Override
